feat: Implement tabbed UI for Vibe Coding projects section
- Add category field ("methodology", "infrastructure", "learning") to each project in PROJECTS array
- Create new ProjectTabs component that renders filterable category tabs
- Create ProjectsPageClient component to manage category filtering state and render filtered projects
- Projects filtered dynamically based on selected category, with "all" showing everything
- Build passes successfully with 56 MDX entries and proper type validation
- Remove 2 malformed auto-generated MDX files (dynamic-context-injection, compound-engineering-philosophy)

The projects page now allows users to filter Vibe Coding portfolio by category using interactive tabs.
